[FIX] BinaryFileAuthHandler теперь работает.
